requirements.txt
# ── Orchestration ──
# apache-airflow==2.8.1 (Not needed for API)

# ── Data Parsing & Scraping ──
pdfplumber==0.10.3
scrapy==2.11.0

# ── Database ──
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# ── Geospatial ──
geopandas==0.14.2
rasterio==1.3.9
rasterstats==0.19.0
shapely==2.0.2
h3==3.7.6

# ── Google Earth Engine ──
earthengine-api==0.1.384

# ── Data Processing ──
pandas==2.1.4
numpy==1.26.3

# ── HTTP & API ──
requests==2.31.0
fastapi==0.109.0
uvicorn==0.27.0
redis==5.0.1
websockets==12.0

# ── ML & Processing ──
scikit-learn==1.4.0
scipy==1.12.0

# ── Configuration ──
python-dotenv==1.0.0

# ── dbt ──
dbt-postgres==1.7.4

# ── Testing ──
pytest==7.4.4
